netbox-event-driven-architectures/requirements.txt
nats-py
python-dotenv
PyGithub
napalm
pythonping
ipaddress
python-nmap
pynetbox
prettytable
